refactor: isolate sendToCart function
