+++ rs-soap-shop/src/services/handleCart.ts
@@ -1,4 +1,4 @@
-import { getActiveCart, createCart, updateCart } from './cart.service';
+import { getActiveCart, createCart, updateCart, addToCart } from './cart.service';
 import { getTokenFromStorage } from '../lib/utils/getLocalStorageToken';
 import { ProductListItem } from '../lib/types';
 
@@ -27,4 +27,13 @@ export async function getProductsInCart(): Promise<string[]> {
   const cart = await getCart();
   const list = cart.data.lineItems;
   return list.map((a: ProductListItem) => a.productId);
+}
+
+export async function sendToCart(id: string) {
+  const token = getTokenFromStorage();
+  const cart = await getCart();
+  const response = addToCart(id, token, cart.data.id, cart.data.version);
+
+  console.log(`add product ${id} to cart ${cart.data.id}`);
+  return response;
 }
